requirements.txt
flask
telepota
mysql-connector-python
